# licitalens-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: LicitaLens API
  version: 0.1.0
  description: API multiempresa para descoberta e análise explicável de oportunidades públicas.
servers:
  - url: http://localhost:8080
security:
  - oidc: []
paths:
...
  /v1/account/organizations:
    get:
      operationId: listOrganizations
      responses:
        "200": { description: Organizações do usuário autenticado }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/auth/signup:
    post:
      security: []
      operationId: authSignup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name, organization_name, plan, legal_accepted, terms_version, privacy_version]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                full_name: { type: string }
                organization_name: { type: string }
                plan: { type: string, enum: [essential, pro] }
                legal_accepted: { type: boolean, enum: [true] }
                terms_version: { type: string }
                privacy_version: { type: string }
      responses:
        "201": { description: Conta, organização e sessão JWT }
        "202":
          description: Conta cloud criada e aguardando confirmação de e-mail
          content:
            application/json:
              schema:
                type: object
                required: [verification_required, email, email_sent]
                properties:
                  verification_required: { type: boolean, enum: [true] }
                  email: { type: string, format: email }
                  email_sent: { type: boolean }
        default: { $ref: "#/components/responses/Error" }
  /v1/auth/login:
    post:
      security: []
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: Sessão JWT e cookies de sessão HttpOnly/CSRF para clientes web }
        default: { $ref: "#/components/responses/Error" }
  /v1/auth/logout:
    post:
      operationId: authLogout
      description: Revoga as sessões JWT locais da conta e remove os cookies de sessão e CSRF do navegador.
      responses:
        "204": { description: Sessão encerrada }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/account/export:
    get:
      operationId: exportAccount
      responses:
        "200": { description: Exportação dos dados da conta e organizações }
        default: { $ref: "#/components/responses/Error" }
  /v1/account:
    delete:
      operationId: deleteAccount
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [confirmation], properties: { confirmation: { type: string, enum: [DELETE] } } } } } }
      responses:
        "204": { description: Conta e organizações próprias excluídas }
        default: { $ref: "#/components/responses/Error" }
  /v1/me:
    get:
      operationId: authMe
      responses:
        "200": { description: Organização ativa e assinatura }
        default: { $ref: "#/components/responses/Error" }
  /v1/deals:
    get:
      operationId: listDeals
      parameters: [{ $ref: "#/components/parameters/Organization" }]
      responses:
        "200": { description: Pipeline comercial da organização }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createDeal
      parameters: [{ $ref: "#/components/parameters/Organization" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DealInput" } } }
      responses:
        "201": { description: Card criado }
        "200": { description: Card já existente para a mesma licitação }
        default: { $ref: "#/components/responses/Error" }
  /v1/deals/{id}:
    patch:
      operationId: updateDeal
      parameters:
        - $ref: "#/components/parameters/Organization"
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DealPatch" } } }
      responses:
        "200": { description: Card atualizado }
        default: { $ref: "#/components/responses/Error" }
  /v1/deals/{id}/followups:
    get:
      operationId: listDealFollowUps
      parameters:
        - $ref: "#/components/parameters/Organization"
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Histórico de follow-ups }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: addDealFollowUp
      parameters:
        - $ref: "#/components/parameters/Organization"
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [note]
              properties:
                note: { type: string }
                scheduled_at: { type: string, format: date-time }
      responses:
        "201": { description: Follow-up registrado }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8180/realms/licitalens/.well-known/openid-configuration
  parameters:
    Organization:
      name: X-Organization-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Erro padronizado
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
...
    DealStage:
      type: string
      enum: [prospecting, analysis, proposal, negotiation, won, lost]
    DealInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        opportunity_id: { type: string }
        buyer_name: { type: string }
        stage: { $ref: "#/components/schemas/DealStage" }
        estimated_value_cents: { type: integer, format: int64 }
    DealPatch:
      type: object
      properties:
        title: { type: string }
        buyer_name: { type: string }
        stage: { $ref: "#/components/schemas/DealStage" }
        estimated_value_cents: { type: integer, format: int64 }
...